Rename ApplicationID to Id for better compatibility with EF
When subclassing an EF model, the ID field of the parent should be "Id" in order to avoid situations where the model has a different name than the parent (all subclasses). EF expects the ID field to be named "Id" or "ClassNameId". Having the Application model's ID field named as anything other than "Id" prevents best compability with EF.